# Unsafe optimisations